fix types

+++ hm_pair.h
@@ -2,6 +2,7 @@
 #define _HM_PAIR_H
 
 #include <stdint.h>
+#include <stdlib.h>
 
 typedef struct HMPair{
 	uint64_t	created;	// 8
@@ -14,8 +15,8 @@ typedef struct HMPair{
 HMPair *hm_pair_create(const char*key, const char*val, unsigned long int expires);
 inline int hm_pair_free(HMPair *pair);
 
-inline const char *hm_pair_getkey(const HMPair *pair, char *buffer, unsigned int len);
-inline const char *hm_pair_getval(const HMPair *pair, char *buffer, unsigned int len);
+inline const char *hm_pair_getkey(const HMPair *pair, char *buffer, size_t len);
+inline const char *hm_pair_getval(const HMPair *pair, char *buffer, size_t len);
 inline int hm_pair_equals(const HMPair *pair, const char *key);
 inline int hm_pair_equalp(const HMPair *pair1, const HMPair *pair2);
